requirements.txt
openai==1.59.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
